Struct gfx_descriptor::DescriptorAllocator

pub struct DescriptorAllocator<B: Backend> { /* fields omitted */ }

Descriptor allocator. Can be used to allocate descriptor sets for any layout.

Implementations

impl<B: Backend> DescriptorAllocator<B>

pub unsafe fn new() -> Self

Create new allocator instance.

Safety

All later operations assume the device is not lost.

pub fn allocate(
&mut self,
device: &B::Device,
layout: &B::DescriptorSetLayout,
layout_counts: &DescriptorCounts,
count: u32,
extend: &mut impl Extend<DescriptorSet<B>>
) -> Result<(), OutOfMemory>

Allocate descriptor set with specified layout. DescriptorCounts must match descriptor numbers of the layout. DescriptorCounts can be constructed from bindings that were used to create layout instance.

pub unsafe fn free(
&mut self,
all_sets: impl IntoIterator<Item = DescriptorSet<B>>
)

Free descriptor sets.

Safety

None of descriptor sets can be referenced in any pending command buffers. All command buffers where at least one of descriptor sets referenced move to invalid state.

pub unsafe fn clear(&mut self, device: &B::Device)

Clear the allocator instance. All sets allocated from this allocator become invalid.

Safety

Assumes none of the allocated blocks will be used from here.

pub fn cleanup(&mut self, device: &B::Device)

Perform cleanup to allow resources reuse.