Crate gfx_descriptor

GPU descriptor allocator

Modules

allocator

counts

Structs

DescriptorAllocator

Descriptor allocator. Can be used to allocate descriptor sets for any layout.

DescriptorCounts

Number of descriptors per type.

DescriptorRangeDesc

Set of descriptors of a specific type.

DescriptorSet

Descriptor set from allocator.

DescriptorSetLayoutBinding

Information about the contents of and in which stages descriptors may be bound to a descriptor set at a certain binding point. Multiple DescriptorSetLayoutBindings are assembled into a DescriptorSetLayout, which is then allocated into a DescriptorSet using a DescriptorPool.

Enums

BufferDescriptorFormat

Format of a buffer.

BufferDescriptorType

Specific type of a buffer.

DescriptorType

The type of a descriptor.

ImageDescriptorType

Specific type of an image descriptor.

Constants

DESCRIPTOR_TYPES

List of all possible descriptor types, with all the possible values the enum variants could have.

DESCRIPTOR_TYPES_COUNT

Length of the DESCRIPTOR_TYPES array.